backend/python-yolo-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ultralytics==8.1.0
Pillow==10.2.0
requests==2.31.0
pydantic==2.5.3
